Move logging.always out of __init__.py
